+++ etna/models/seasonal_ma.py
@@ -1,10 +1,16 @@
 import warnings
+from typing import Dict
 
 import numpy as np
 import pandas as pd
 
+from etna import SETTINGS
 from etna.datasets import TSDataset
 from etna.models.base import NonPredictionIntervalContextRequiredAbstractModel
+
+if SETTINGS.auto_required:
+    from optuna.distributions import BaseDistribution
+    from optuna.distributions import IntUniformDistribution
 
 
 class SeasonalMovingAverageModel(
@@ -190,6 +196,18 @@ class SeasonalMovingAverageModel(
         new_df = self._predict(df=df, prediction_size=prediction_size)
         ts.df = new_df
         return ts
+
+    def params_to_tune(self) -> Dict[str, "BaseDistribution"]:
+        """Get default grid for tuning hyperparameters.
+
+        This grid doesn't tune ``seasonality`` parameter. It expected to be set by the user.
+
+        Returns
+        -------
+        :
+            Grid to tune.
+        """
+        return {"window": IntUniformDistribution(low=1, high=10, step=1)}
 
 
 __all__ = ["SeasonalMovingAverageModel"]